service discovery added

+++ ara/com/network_binding/SomeIp/servicediscovery/lib/include/entry.hpp
@@ -11,7 +11,7 @@
 class entry
 {
 protected:
-uint8_t type;
+uint16_t type;
 uint8_t index_1st_options; //unique pointer
 uint8_t index_2nd_options;
 uint8_t no_1st_options;
@@ -22,7 +22,7 @@ uint16_t Service_ID;
 uint16_t Instance_ID;
 entry();
 public:
-virtual void setType (uint8_t type)=0;
+virtual void setType (uint16_t type)=0;
 virtual void setIndexFirstOption(uint8_t index_1st_options)=0;
 virtual void setIndexSecondOption(uint8_t index_2nd_options)=0;
 virtual void setMajorVersion(uint16_t major_version)=0;
@@ -32,7 +32,7 @@ virtual void setTimeToLive(uint32_t ttl)=0;
 virtual uint32_t getTimeToLIve(void)=0;
 virtual void setServiceID(uint16_t Service_ID)=0;
 virtual void setInstanceID(uint16_t Instance_ID)=0;
-virtual uint8_t getType(void)=0;
+virtual uint16_t getType(void)=0;
 virtual uint16_t getServiceID(void)=0;
 virtual uint16_t getInstanceID(void)=0;
 
@@ -47,7 +47,7 @@ class service_entry:public entry
     public:
     service_entry();
     service_entry(uint16_t serviceID,uint16_t InstanceID);  //check
-     void setType (uint8_t type);
+     void setType (uint16_t type);
      void setIndexFirstOption(uint8_t index_1st_options);
      void setIndexSecondOption(uint8_t index_2nd_options);
      void setMajorVersion(uint16_t major_version);
@@ -60,7 +60,7 @@ class service_entry:public entry
      uint16_t getServiceID(void);
      uint16_t getInstanceID(void);
      uint32_t getTimeToLIve(void);
-     uint8_t getType(void);
+     uint16_t getType(void);
     static service_entry *create_find_service_entry(uint16_t Service_ID,uint16_t Instance_ID);  //changeddd
     static service_entry * create_offer_service_entry(uint16_t Service_ID,uint16_t Instance_ID);
     static service_entry * create_stop_offer_service_entry(uint16_t Service_ID,uint16_t Instance_ID);
